+++ CHANGELOG.md
@@ -10,3 +10,4 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 ## 0.3.0
 
 - Migrated the pipeline from the .NET version of the package.
+- Ported the code from the .NET version of the package to dart.